requirements.txt
langchain
langchain_core
# google-generativeai
# google.generativeai
langchain_google_genai
langserve
fastapi
uvicorn
python-dotenv